thumbnailing: Move resizing functions into zerver.lib.thumbnail.
