+++ zerver/tests/test_upload_local.py
@@ -10,6 +10,7 @@ import zerver.lib.upload
 from zerver.lib.avatar_hash import user_avatar_path
 from zerver.lib.test_classes import UploadSerializeMixin, ZulipTestCase
 from zerver.lib.test_helpers import get_test_image_file, read_test_image_file
+from zerver.lib.thumbnail import DEFAULT_EMOJI_SIZE, MEDIUM_AVATAR_SIZE, resize_avatar
 from zerver.lib.upload import (
     all_message_attachments,
     delete_export_tarball,
@@ -20,7 +21,6 @@ from zerver.lib.upload import (
     upload_export_tarball,
     upload_message_attachment,
 )
-from zerver.lib.upload.base import DEFAULT_EMOJI_SIZE, MEDIUM_AVATAR_SIZE, resize_avatar
 from zerver.lib.upload.local import write_local_file
 from zerver.models import Attachment, RealmEmoji
 from zerver.models.realms import get_realm
